refactor: extract SyncManager from LocationForegroundService
Move sync queue processing, batch uploads, retry logic, and backoff
strategy into a dedicated SyncManager class to improve separation of
concerns and reduce service file size
